refactor(agent): move interfaces to types/ folder for architectural compliance
- Move AgentContentViewProps to types/AgentContentView.ts
- Move AuthPending to types/AuthPending.ts
- Create types/index.ts for centralized exports
- Update all imports across agent module
