+++ src/renderer/components/states/agent/types/AuthPending.ts
@@ -1,0 +1,34 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file AuthPending.ts
+ * @description Agent 授权待处理状态类型定义
+ * @author 鸡哥
+ */
+
+/** Agent 授权待处理状态 */
+export interface AuthPending {
+  type: 'web' | 'tool';
+  requestId: string;
+  description: string;
+  tool?: string;
+  argumentsPayload?: Record<string, unknown>;
+}
